requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.4
python-dotenv>=1.0.1
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
aiohttp>=3.11.0
requests>=2.32.3
alembic>=1.14.0
itsdangerous>=2.2.0
email-validator>=2.2.0
pydantic-settings>=2.6.0
python-multipart>=0.0.18
aiofiles>=24.1.0
apscheduler>=3.10.4
pydantic>=2.0.0
minio>=7.2.0
python-dateutil>=2.8.2
cryptography>=41.0.0
psycopg2-binary>=2.9.9
paramiko>=3.4.0

